pyproject.toml
[build-system]
# https://setuptools.readthedocs.io/en/latest/setuptools.html#setup-cfg-only-projects
requires = [
  "setuptools == 68.2.2",
  "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "SSLSocket"
authors = [
    {name = "KhanhNguyen9872", email = "khanhvuinguyenvan800983609@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.6"
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]

dependencies = [
    "pyOpenSSL"
]